Started new example ZhangExampleFDM based on TumorHeatEqnFDM
Started new example ZhangExampleFDM which is based on files from TumorHeatEqnFDM.
ZhangExampleFDM will be used to implement the example from Zhang 2008 paper.
The example will be used to validate discretization of the Pennes Bioheat Equation with a analytical solution presented in the paper.
Functions and the documentation are currently empty.
Example was added to all the make and configure files.