Updated TumorHeatEqnFDM example
Removed setting perfusion rate (omega) in evalInner.
One additional vector will be used to distinguish between brain and tumor region.
This vector will either be set in initInner or by the python script and then read from file.
In updateInner and and updateBorder perfusion rate and metabolism heat will be set to value for brain or for tumor.
Needed to have at least one vector which is a knownDf.