Added two parameters for convergence check
Added checkConvFirstAtIter to specify the first the iteration when the convergence check should be computed for the first time.
This parameter is set via environment variable.
Default value is 1.
Added checkConvAtEveryNIter to specify the number of iterations between two convergence checks.
This parameter is set via environment variable.
Default value is 1.
With both parameters one can reduce the number of convergence checks.