Fixed error if only one unknown datafield is used
ScaFES did not run correctly if only one unknown datafield was used and no known datafield.
This happened because some values were called which were not initialized because there is no known datafield vector.
Added if statement which is only true if at least one known datafield exists or compute error is true for the unknown datafield.