Updated plot scaling results
Changed behaviour of deleting highest and lowest value.
Index of values which are to be deleted are determined by highest and lowest of value of the total runtime.
This set of index will be used for all arrays.
I.e. in update, write, ... the values which are the highest and lowest in total runtime will be deleted.
Therefore it is not necessary that the highest or lowest value of update, write, ... are deleted.